Move components and data to seperate files
