+++ src/data.ts
@@ -1,0 +1,19 @@
+export const sampleTodos = [
+    {
+        content: 'Welcome to Toododo!',
+        isCompleted: true,
+    },
+    {
+        content: '"Enter" in any todo to add a new todo below it',
+        isCompleted: false,
+    },
+    {
+        content: 'Click + to add a todo in the bottom of the list',
+        isCompleted: false,
+    },
+    {
+        content: 'Click x to remove a todo or remove all in the bottom',
+        isCompleted: false,
+    }
+
+];
